backend/requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# QuizGen Platform — Python Dependencies
# FILE : requirements.txt

# Install:  pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────────────────────

# ── Web Framework ─────────────────────────────────────────────────────────────
Flask==3.0.3
Flask-CORS==4.0.1
Werkzeug==3.0.3

# ── Database ──────────────────────────────────────────────────────────────────
flask-pymongo==2.3.0
pymongo[srv]==4.7.3
dnspython==2.6.1

# ── Authentication ────────────────────────────────────────────────────────────
PyJWT==2.8.0

# ── Environment & HTTP ────────────────────────────────────────────────────────
python-dotenv==1.0.1
requests==2.32.3 

# ── Production Server ─────────────────────────────────────────────────────────
gunicorn==22.0.0
certifi==2024.7.4 

# ── Image Handling ────────────────────────────────────────────────────────────
Pillow==10.4.0